TS-Doc comments, moved type definitions around
